+++ src/contracts/smartContract.js
