+++ src/lib/hooks/usePDNBalance.js
@@ -1,7 +1,7 @@
 import { useMoralis, useWeb3ExecuteFunction } from "react-moralis";
-import { erc20Options } from "src/abis";
+import { erc20Options } from "src/contracts/sc-options";
 
-import SMART_CONTRACT_FUNCTIONS from "src/smartContract";
+import SMART_CONTRACT_FUNCTIONS from "src/contracts/smartContract";
 
 function usePDNBalance({ account: userAccount } = {}) {
   const { Moralis, user } = useMoralis();
