+++ src/lib/hooks/usePDNSymbol.js
@@ -1,7 +1,7 @@
 import { useWeb3ExecuteFunction } from "react-moralis";
-import { erc20Options } from "src/abis";
+import { erc20Options } from "src/contracts/sc-options";
 
-import SMART_CONTRACT_FUNCTIONS from "src/smartContract";
+import SMART_CONTRACT_FUNCTIONS from "src/contracts/smartContract";
 
 function usePDNSymbol() {
   const options = erc20Options(null, SMART_CONTRACT_FUNCTIONS.SYMBOL);
