+++ src/pages/Accessibility.js
@@ -1,18 +1,18 @@
 // components
-import Page from '../components/Page';
-import SMART_CONTRACT_FUNCTIONS from 'src/smartContract';
-import FunctionsMenu from 'src/components/FunctionsMenu';
+import Page from "../components/Page";
+import SMART_CONTRACT_FUNCTIONS from "src/contracts/smartContract";
+import FunctionsMenu from "src/components/FunctionsMenu";
 
 // ----------------------------------------------------------------------
 
 const availableFunctions = [
   {
-    name: "Freeze DAO", 
+    name: "Freeze DAO",
     functionName: SMART_CONTRACT_FUNCTIONS.FREEZE_DAO,
     args: {},
     event: SMART_CONTRACT_FUNCTIONS.FREEZE_DAO,
-  }, 
-]
+  },
+];
 
 export default function Accessibility() {
   return (
